+++ rrule/examples/manual_rrule.rs
@@ -3,15 +3,15 @@
 //! Create an [`RRule`] object.
 
 use chrono::{Datelike, TimeZone, Timelike};
-use chrono_tz::UTC;
-use rrule::{Frequency, RRule};
+use rrule::{Frequency, RRule, Tz};
 
 fn main() {
     // Build an RRuleSet that starts first day in 2020 at 9:00AM and occurs daily 5 times
+    let start_date = Tz::UTC.ymd(2020, 1, 1).and_hms(9, 0, 0);
     let rrule_set = RRule::default()
         .count(5)
         .freq(Frequency::Daily)
-        .build(UTC.ymd(2020, 1, 1).and_hms(9, 0, 0))
+        .build(start_date)
         .expect("RRule invalid");
 
     let recurrences = rrule_set.all_unchecked();