+++ rrule/src/core/datetime.rs
@@ -1,5 +1,5 @@
+use super::timezone::Tz;
 use chrono::{Datelike, Duration, NaiveTime, Timelike};
-use chrono_tz::Tz;
 
 pub(crate) type DateTime = chrono::DateTime<Tz>;
 
@@ -36,11 +36,17 @@ pub(crate) fn datetime_to_ical_format(dt: &DateTime) -> String {
     let mut tz_prefix = String::new();
     let mut tz_postfix = String::new();
     let tz = dt.timezone();
-    if tz == Tz::UTC {
-        tz_postfix = "Z".to_string();
-    } else {
-        tz_prefix = format!(";TZID={}", tz.name());
-    };
+    match tz {
+        Tz::Local(_) => {}
+        Tz::Tz(tz) => match tz {
+            chrono_tz::UTC => {
+                tz_postfix = "Z".to_string();
+            }
+            tz => {
+                tz_prefix = format!(";TZID={}", tz.name());
+            }
+        },
+    }
 
     let dt = dt.format("%Y%m%dT%H%M%S");
     format!("{}:{}{}", tz_prefix, dt, tz_postfix)