+++ rrule/src/validator/validate_rrule.rs
@@ -1,7 +1,7 @@
 use std::ops::RangeInclusive;
 
 use crate::core::DateTime;
-use crate::{Frequency, NWeekday, RRule, Unvalidated};
+use crate::{Frequency, NWeekday, RRule, Tz, Unvalidated};
 
 use super::ValidationError;
 
@@ -45,13 +45,44 @@ pub(crate) fn validate_rrule_forced(
 }
 
 // Until:
+// - Timezones are correctly synced as specified in the RFC
 // - Value should be later than `dt_start`.
 fn validate_until(rrule: &RRule<Unvalidated>, dt_start: &DateTime) -> Result<(), ValidationError> {
     match rrule.until {
-        Some(until) if until < *dt_start => Err(ValidationError::UntilBeforeStart {
-            until: until.to_rfc3339(),
-            dt_start: dt_start.to_rfc3339(),
-        }),
+        Some(until) => {
+            match dt_start.timezone() {
+                Tz::Local(_) => {
+                    let allowed_timezones = vec![Tz::LOCAL, Tz::UTC];
+                    if !allowed_timezones.contains(&until.timezone()) {
+                        return Err(ValidationError::DtStartUntilMismatchTimezone {
+                            dt_start_tz: dt_start.timezone().name().into(),
+                            until_tz: until.timezone().name().into(),
+                            expected: allowed_timezones
+                                .into_iter()
+                                .map(|tz| tz.name().into())
+                                .collect(),
+                        });
+                    }
+                }
+                Tz::Tz(_) => {
+                    if until.timezone() != Tz::UTC {
+                        return Err(ValidationError::DtStartUntilMismatchTimezone {
+                            dt_start_tz: dt_start.timezone().name().into(),
+                            until_tz: until.timezone().name().into(),
+                            expected: vec!["UTC".into()],
+                        });
+                    }
+                }
+            }
+
+            if until < *dt_start {
+                return Err(ValidationError::UntilBeforeStart {
+                    until: until.to_rfc3339(),
+                    dt_start: dt_start.to_rfc3339(),
+                });
+            }
+            Ok(())
+        }
         _ => Ok(()),
     }
 }
@@ -328,9 +359,12 @@ fn validate_not_equal_for_vec<T: PartialEq<T> + ToString>(
 #[cfg(test)]
 mod tests {
     use chrono::TimeZone;
-    use chrono_tz::UTC;
+
+    use crate::core::Tz;
 
     use super::*;
+
+    const UTC: Tz = Tz::UTC;
 
     #[test]
     fn rejects_by_set_pos_without_byxxx_rule() {
@@ -548,5 +582,72 @@ mod tests {
                 dt_start: dt_start.to_rfc3339()
             }
         );
+    }
+
+    #[test]
+    fn allows_until_with_compatible_timezone() {
+        fn t(start_tz: Tz, until_tz: Tz) -> (DateTime, DateTime) {
+            (
+                start_tz
+                    .ymd_opt(2020, 1, 1)
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .into(),
+                until_tz
+                    .ymd_opt(2020, 1, 1)
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .into(),
+            )
+        }
+
+        let tests = [t(Tz::LOCAL, Tz::LOCAL), t(Tz::LOCAL, UTC), t(UTC, UTC)];
+
+        for (start_date, until) in tests {
+            let rrule = RRule {
+                until: Some(until),
+                ..Default::default()
+            };
+            let res = validate_rrule_forced(&rrule, &start_date);
+            assert!(res.is_ok());
+        }
+    }
+
+    #[test]
+    fn rejects_until_with_incompatible_timezone() {
+        fn t(start_tz: Tz, until_tz: Tz) -> (DateTime, DateTime) {
+            (
+                start_tz
+                    .ymd_opt(2020, 1, 1)
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .into(),
+                until_tz
+                    .ymd_opt(2020, 1, 1)
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .into(),
+            )
+        }
+
+        let tests = [
+            t(Tz::UTC, Tz::LOCAL),
+            t(Tz::Europe__Berlin, Tz::LOCAL),
+            t(Tz::LOCAL, Tz::Europe__Berlin),
+        ];
+
+        for (start_date, until) in tests {
+            let rrule = RRule {
+                until: Some(until),
+                ..Default::default()
+            };
+            let res = validate_rrule_forced(&rrule, &start_date);
+            assert!(res.is_err());
+            let err = res.unwrap_err();
+            assert!(matches!(
+                err,
+                ValidationError::DtStartUntilMismatchTimezone { .. }
+            ));
+        }
     }
 }